MAINT: make space for expanding more solvers
