+++ a_package/solvers/__init__.py
@@ -1,0 +1,2 @@
+from .contact import RigidContactSolver
+from .phase import PhaseSolver
